refractoring and DRYing
