+++ aileen/box/urls.py
@@ -1,10 +1,9 @@
 from django.conf.urls import url
 
-from box import views, api
+from box import views
 
 app_name = "box"
 
 urlpatterns = [
     url(r"^$", views.dashboard, name="dashboard"),
-    url(r"^api/devices/", api.devices, name="devices"),
 ]
