+++ aileen/server/templates/server/box.html
@@ -103,11 +103,11 @@
 
   <!-- charts -->
   <script type="text/javascript">
-  const aileen_data = `{% url "server:devices_by_box_id" box_id %}`;
+  const aileen_data = `{% url "data:devices_by_box_id" box_id %}`;
   </script>
-  <script src="{% static 'server/js/box-heatmapchart.js' %}"></script>
-  <script src="{% static 'server/js/box-radialbarchart.js' %}"></script>
-  <script src="{% static 'server/js/box-main.js' %}"></script>
+  <script src="{% static 'data/js/heatmapchart.js' %}"></script>
+  <script src="{% static 'data/js/radialbarchart.js' %}"></script>
+  <script src="{% static 'server/js/d3-main.js' %}"></script>
   <script type="text/javascript">
   var dataurl = '{% url "server:aileen_box_location" %}';
   window.addEventListener("map:init", function (event) {
